explor_converg_eval/requirements.txt
aiohttp>=3.9.0
asyncio
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pydantic>=2.0.0
instructor>=1.0.0
openai>=1.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
aiosqlite>=0.19.0
pyyaml>=6.0.0
